Simplify model

+++ src/Builders/PassBuilder.php
@@ -4,20 +4,38 @@ namespace Spatie\LaravelMobilePass\Builders;
 
 use Illuminate\Support\Collection;
 use PKPass\PKPass;
+use Spatie\LaravelMobilePass\Entities\Colour;
 use Spatie\LaravelMobilePass\Entities\FieldContent;
 use Spatie\LaravelMobilePass\Entities\Image;
 use Spatie\LaravelMobilePass\Entities\Price;
 use Spatie\LaravelMobilePass\Entities\WifiNetwork;
 use Spatie\LaravelMobilePass\Enums\PassType;
+use Spatie\LaravelMobilePass\Models\MobilePass;
 use Spatie\LaravelMobilePass\Validators\PassValidator;
 
 abstract class PassBuilder
 {
+    protected ?array $data = null;
+
     protected PassType $type;
 
     protected ?string $serialNumber = null;
 
     protected ?string $organisationName = null;
+
+    protected ?string $passTypeIdentifier = null;
+
+    protected ?string $authenticationToken = null;
+
+    protected ?string $teamIdentifier = null;
+
+    protected ?Colour $backgroundColour = null;
+
+    protected ?Colour $foregroundColour = null;
+
+    protected ?Colour $labelColour = null;
+
+    protected ?PassType $passType = null;
 
     protected ?string $description = null;
 
@@ -39,9 +57,48 @@ abstract class PassBuilder
 
     abstract protected static function validator(): PassValidator;
 
-    public static function make(): static
+    public static function make(array $data = [], array $images = [], ?MobilePass $model = null): static
     {
-        return new static;
+        return new static($data, $images, $model);
+    }
+
+    public function __construct(array $data = [], array $images = [], protected ?MobilePass $model = null)
+    {
+        $this->data = $data;
+        $this->images = $images;
+
+        $this->uncompileContent();
+    }
+
+    protected function uncompileContent()
+    {
+        $this->organisationName = $this->data['organisationName'] ?? null;
+        $this->passTypeIdentifier = $this->data['passTypeIdentifier'] ?? null;
+        $this->authenticationToken = $this->data['authenticationToken'] ?? null;
+        $this->teamIdentifier = $this->data['teamIdentifier'] ?? null;
+        $this->description = $this->data['description'] ?? null;
+        $this->backgroundColour = Colour::makeFromRgbString($this->data['backgroundColor'] ?? null);
+        $this->foregroundColour = Colour::makeFromRgbString($this->data['foregroundColor'] ?? null);
+        $this->labelColour = Colour::makeFromRgbString($this->data['labelColor'] ?? null);
+        $this->passType = PassType::tryFrom($this->data['userInfo']['passType'] ?? PassType::Generic->value);
+
+        // $model->passImages = array_map(fn ($image) => Image::fromArray($image), $model->images);
+        // $model->barcodes = array_map(fn ($barcode) => Barcode::fromArray($barcode), $model->content['barcodes'] ?? []);
+
+        $this->uncompileFieldSet('headerFields');
+        $this->uncompileFieldSet('primaryFields');
+        $this->uncompileFieldSet('secondaryFields');
+        $this->uncompileFieldSet('auxiliaryFields');
+        $this->uncompileFieldSet('backFields');
+    }
+
+    protected function uncompileFieldSet(string $fieldSetName)
+    {
+        $this->$fieldSetName = collect();
+
+        foreach ($this->data[$this->passType->value][$fieldSetName] ?? [] as $field) {
+            $this->$fieldSetName[$field['key']] = FieldContent::fromArray($field);
+        }
     }
 
     public function setLogoImage(Image $image): self
@@ -82,6 +139,34 @@ abstract class PassBuilder
     public function setHeaderFields(FieldContent ...$headerField): self
     {
         $this->headerFields = collect($headerField);
+
+        return $this;
+    }
+
+    public function setBackFields(FieldContent ...$backField): self
+    {
+        $this->backFields = collect($backField);
+
+        return $this;
+    }
+
+    public function updateField(string $key, string $value)
+    {
+        $fieldTypes = [
+            'headerFields',
+            'primaryFields',
+            'secondaryFields',
+            'auxiliaryFields',
+            'backFields',
+        ];
+
+        foreach ($fieldTypes as $fieldType) {
+            $this->$fieldType = $this->$fieldType->map(function ($field) use ($key, $value) {
+                if ($field->key === $key) {
+                    $field->value = $value;
+                }
+            });
+        }
 
         return $this;
     }
@@ -134,7 +219,7 @@ abstract class PassBuilder
 
     protected function compileData(): array
     {
-        return [
+        return array_merge($this->data ?? [], array_filter([
             'formatVersion' => 1,
             'organizationName' => $this->organisationName,
             'passTypeIdentifier' => config('mobile-pass.type_identifier'),
@@ -143,7 +228,7 @@ abstract class PassBuilder
             'teamIdentifier' => config('mobile-pass.team_identifier'),
             'description' => $this->description,
             'semantics' => $this->compileSemantics(),
-        ];
+        ]));
     }
 
     protected function addImagesToFile(PKPass $pkPass): PKPass
@@ -151,6 +236,10 @@ abstract class PassBuilder
         foreach ($this->images as $filename => $image) {
             // The $image Image entity could contain up to three
             // images in different resolutions.
+
+            if (! $image instanceof Image) {
+                $image = Image::make($image['x1Path'], $image['x2Path'], $image['x3Path']);
+            }
 
             if ($image->x1Path) {
                 $pkPass->addFile($image->x1Path, "$filename.png");
@@ -168,17 +257,19 @@ abstract class PassBuilder
         return $pkPass;
     }
 
-    protected static function getCertificatePath(): string
+    public static function getCertificatePath(): string
     {
         if (! empty(config('mobile-pass.apple.certificate_contents'))) {
-            $path = __DIR__.'/../../tmp/Cert.p12';
+            $path = sys_get_temp_dir().'/LaravelMobilePass.p12';
 
-            file_put_contents(
-                $path,
-                base64_decode(
-                    config('mobile-pass.apple.certificate_contents')
-                )
-            );
+            if (! file_exists($path)) {
+                file_put_contents(
+                    $path,
+                    base64_decode(
+                        config('mobile-pass.apple.certificate_contents')
+                    )
+                );
+            }
 
             return $path;
         }
@@ -186,7 +277,30 @@ abstract class PassBuilder
         return config('mobile-pass.apple.certificate_path');
     }
 
-    public function generate()
+    public static function getCertificatePassword(): string
+    {
+        return config('mobile-pass.apple.certificate_password');
+    }
+
+    public function save(): MobilePass
+    {
+        if ($this->model) {
+            $this->model->update([
+                'content' => $this->data(),
+                'images' => $this->images,
+            ]);
+
+            return $this->model;
+        }
+
+        return MobilePass::create([
+            'builder_class' => static::class,
+            'content' => $this->data(),
+            'images' => $this->images,
+        ]);
+    }
+
+    public function data(): array
     {
         if (empty($this->organisationName)) {
             $this->setOrganisationName(
@@ -208,12 +322,18 @@ abstract class PassBuilder
         // The icon image is always required.
         // TODO: validate this.
 
+        return $data;
+    }
+
+    public function generate()
+    {
+
         $pkPass = new PKPass(
             self::getCertificatePath(),
-            config('mobile-pass.apple.certificate_password'),
+            self::getCertificatePassword(),
         );
 
-        $pkPass->setData($data);
+        $pkPass->setData($this->data());
 
         $this->addImagesToFile($pkPass);
